refactor: money에 화폐 종류 추가

+++ src/main/java/shopping/dto/web/response/OrderItemResponse.java
@@ -22,7 +22,7 @@ public class OrderItemResponse {
                 orderItem.getId(),
                 orderItem.getProduct().getName(),
                 orderItem.getProduct().getImageUrl(),
-                orderItem.getProduct().getPrice().getPrice(),
+                orderItem.getProduct().getPrice().getPrice().longValue(),
                 orderItem.getQuantity().getQuantity()
         );
     }
